chore: Move Separator component to UI directory
