+++ packages/webapp-libs/webapp-core/src/components/pageHeadline/pageHeadline.component.tsx
@@ -1,8 +1,8 @@
 import { ReactNode } from 'react';
 
 import { BackButton } from '../buttons/backButton';
-import { Separator } from '../separator';
 import { H3, Paragraph } from '../typography';
+import { Separator } from '../ui/separator';
 
 export type PageHeadlineProps = {
   hasBackButton?: boolean;
