--- packages/webapp-libs/webapp-core/src/components/separator/index.ts
@@ -1,1 +1,0 @@
-export * from './separator.component';
