+++ packages/webapp-libs/webapp-core/src/components/ui/separator.tsx
@@ -1,7 +1,7 @@
 import * as SeparatorPrimitive from '@radix-ui/react-separator';
 import * as React from 'react';
 
-import { cn } from '../../lib/utils';
+import { cn } from '@sb/webapp-core/lib/utils';
 
 const Separator = React.forwardRef<
   React.ElementRef<typeof SeparatorPrimitive.Root>,
@@ -9,7 +9,6 @@ const Separator = React.forwardRef<
 >(({ className, orientation = 'horizontal', decorative = true, ...props }, ref) => (
   <SeparatorPrimitive.Root
     ref={ref}
-    data-testid="Separator"
     decorative={decorative}
     orientation={orientation}
     className={cn('bg-border shrink-0', orientation === 'horizontal' ? 'h-[1px] w-full' : 'h-full w-[1px]', className)}
