+++ packages/webapp/src/shared/components/auth/addTwoFactorAuth/addTwoFactorAuth.component.tsx
@@ -3,8 +3,8 @@ import { useApiForm } from '@sb/webapp-api-client/hooks';
 import { useCommonQuery } from '@sb/webapp-api-client/providers';
 import { Button, ButtonVariant } from '@sb/webapp-core/components/buttons';
 import { Input } from '@sb/webapp-core/components/forms';
-import { Separator } from '@sb/webapp-core/components/separator';
 import { H1, H4, Paragraph, Small } from '@sb/webapp-core/components/typography';
+import { Separator } from '@sb/webapp-core/components/ui/separator';
 import { trackEvent } from '@sb/webapp-core/services/analytics';
 import { useToast } from '@sb/webapp-core/toast/useToast';
 import * as QRCode from 'qrcode';
